# backend/contract/openapi.yaml
openapi: 3.0.3
info:
  title: Beatris API
  version: 1.0.0
  description: API веб-приложения медицинского центра эстетической медицины Beatris
servers:
  - url: /api/v1
    description: Основной относительный адрес API
...
paths:
...
  /users/me:
    get:
      tags:
        - Users
      summary: Получить профиль пользователя
      operationId: getMyProfile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Профиль пользователя успешно получен
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"

    patch:
      tags:
        - Users
      summary: Изменить профиль пользователя
      operationId: updateMyProfile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Профиль пользователя успешно изменён
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /specialists:
    get:
      tags:
        - Specialists
      summary: Получить список специалистов
      operationId: getSpecialists
      parameters:
        - $ref: "#/components/parameters/Search"
        - name: serviceId
          in: query
          description: Фильтр по идентификатору услуги
          required: false
          schema:
            type: string
            format: uuid
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Список специалистов успешно получен
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpecialistListResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /specialists/{id}:
    get:
      tags:
        - Specialists
      summary: Получить специалиста по id
      operationId: getSpecialistById
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Специалист успешно получен
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Specialist"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /wishlist:
    get:
      tags:
        - Wishlist
      summary: Получить избранные товары
      operationId: getWishlist
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Избранные товары успешно получены
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WishlistItem"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /wishlist/items:
    post:
      tags:
        - Wishlist
      summary: Добавить товар в избранное
      operationId: addWishlistItem
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddWishlistItemRequest"
      responses:
        "201":
          description: Товар успешно добавлен в избранное
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WishlistItem"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /wishlist/items/{productId}:
    delete:
      tags:
        - Wishlist
      summary: Удалить товар из избранного
      operationId: deleteWishlistItem
      security:
        - bearerAuth: []
      parameters:
        - name: productId
          in: path
          description: Идентификатор товара
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Товар успешно удалён из избранного
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token, передаваемый в заголовке Authorization
    refreshCookie:
      type: apiKey
      in: cookie
      name: refreshToken
      description: Refresh token, хранящийся в защищённой HttpOnly cookie
  parameters:
    Id:
      name: id
      in: path
      description: Уникальный идентификатор ресурса
      required: true
      schema:
        type: string
        format: uuid
        example: 550e8400-e29b-41d4-a716-446655440000
...
    Search:
      name: search
      in: query
      description: Поисковая строка
      required: false
      schema:
        type: string
        example: акне
    Page:
      name: page
      in: query
      description: Номер страницы
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    Limit:
      name: limit
      in: query
      description: Количество элементов на странице
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
    BadRequest:
      description: Некорректный запрос
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            statusCode: 400
            message: Некорректные данные запроса
            error: Bad Request
    Unauthorized:
      description: Пользователь не авторизован
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            statusCode: 401
            message: Необходимо выполнить вход в систему
            error: Unauthorized
...
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            statusCode: 404
            message: Запрашиваемый ресурс не найден
            error: Not Found
    ValidationError:
      description: Ошибка валидации данных
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationErrorResponse"
          example:
            statusCode: 422
            message:
              - phone должен быть указан
              - serviceId должен быть UUID
            error: Unprocessable Entity
    InternalServerError:
      description: Внутренняя ошибка сервера
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            statusCode: 500
            message: Внутренняя ошибка сервера
            error: Internal Server Error
  schemas:
    Role:
      type: string
      enum:
        - patient
        - doctor
        - admin
      example: patient
...
    PaginationMeta:
      type: object
      required:
        - page
        - limit
        - total
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 120
    ErrorResponse:
      type: object
      required:
        - statusCode
        - message
        - error
      properties:
        statusCode:
          type: integer
          example: 400
        message:
          type: string
          example: Некорректные данные запроса
        error:
          type: string
          example: Bad Request
    ValidationErrorResponse:
      type: object
      required:
        - statusCode
        - message
        - error
      properties:
        statusCode:
          type: integer
          example: 422
        message:
          type: array
          items:
            type: string
          example:
            - phone должен быть указан
            - serviceId должен быть UUID
        error:
          type: string
          example: Unprocessable Entity
    User:
      type: object
      description: Пользователь системы
      required:
        - id
        - fullName
        - phone
        - role
      properties:
        id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        fullName:
          type: string
          example: Айгуль Нурланова
        phone:
          type: string
          example: "+77001234567"
        email:
          type: string
          format: email
          nullable: true
          example: user@example.com
        role:
          $ref: "#/components/schemas/Role"
        createdAt:
          type: string
          format: date-time
          example: "2026-05-20T10:00:00Z"
...
    UpdateProfileRequest:
      type: object
      properties:
        fullName:
          type: string
          example: Айгуль Нурланова
        phone:
          type: string
          example: "+77001234567"
        email:
          type: string
          format: email
          nullable: true
          example: user@example.com
...
    Specialist:
      type: object
      description: Специалист медицинского центра
      required:
        - id
        - fullName
        - position
        - specialization
      properties:
        id:
          type: string
          format: uuid
          example: 4bd38fe7-6d07-43c6-81c1-47e7e1279356
        fullName:
          type: string
          example: Иванова Анна Сергеевна
        position:
          type: string
          example: Врач-косметолог
        specialization:
          type: string
          example: Инъекционная косметология
        experienceYears:
          type: integer
          example: 8
        photoUrl:
          type: string
          nullable: true
          example: /uploads/specialists/ivanova.jpg
        serviceIds:
          type: array
          items:
            type: string
            format: uuid
...
    SpecialistListResponse:
      type: object
      required:
        - items
        - meta
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Specialist"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
...
    ProductCategory:
      type: object
      required:
        - id
        - title
        - slug
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          example: Средства для ухода
        slug:
          type: string
          example: skincare
    Product:
      type: object
      description: Товар интернет-магазина
      required:
        - id
        - title
        - slug
        - description
        - price
        - category
        - stock
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          example: Крем для восстановления кожи
        slug:
          type: string
          example: vosstanavlivayushchiy-krem
        description:
          type: string
          example: Профессиональное средство для домашнего ухода
        price:
          type: number
          minimum: 0
          example: 18000
        category:
          $ref: "#/components/schemas/ProductCategory"
        imageUrl:
          type: string
          nullable: true
          example: /uploads/products/cream.jpg
        stock:
          type: integer
          minimum: 0
          example: 12
...
    AddWishlistItemRequest:
      type: object
      required:
        - productId
      properties:
        productId:
          type: string
          format: uuid
    WishlistItem:
      type: object
      required:
        - id
        - product
      properties:
        id:
          type: string
          format: uuid
        product:
          $ref: "#/components/schemas/Product"
        createdAt:
          type: string
          format: date-time
...